further break down mmu_read and mmu_write
I think that I'm going to have to take this one block of memory at a time - tackling both read and write simultaneously.
That way I can keep the underlying memory management on one side of the boundary or the other.
Start by making a 'switch' statement on addr and then sending else down the ccall black hole.