clock, oam, and the start of video_update
as far as I can tell the types work. these functions are quite closely coupled. so it is tricky to identify if they are working until they are all ported. opting to just byte the bullet and see if it I can get it working. worst case I go back to main branch and try a different strategy.